# emmaus_ui/src/app/services/models/staff-profile-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface StaffProfileResponse {
  bio?: string;
  createdAt?: string;
  department?: string;
  description?: string;
  displayOrder?: number;
  email?: string;
  firstName?: string;
  lastName?: string;
  linkedinUrl?: string;
  phone?: string;
  position?: string;
  profilePic?: string;
  publicId?: string;
}