# emmaus_ui/src/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/staff-profiles/update/{public-id}": {
            "post": {
                "tags": [
                    "staff-profiles"
                ],
                "operationId": "updateProfile",
                "parameters": [
                    {
                        "name": "public-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatStaffProfileRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/staff-profiles/create": {
            "post": {
                "tags": [
                    "staff-profiles"
                ],
                "operationId": "createStaffProfile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatStaffProfileRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/staff-profiles/pages": {
            "get": {
                "tags": [
                    "staff-profiles"
                ],
                "operationId": "getPagesOfStaffProfile",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "position",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "department",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseStaffProfileResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/staff-profiles/delete/{public-id}": {
            "delete": {
                "tags": [
                    "staff-profiles"
                ],
                "operationId": "deleteProfile",
                "parameters": [
                    {
                        "name": "public-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreatStaffProfileRequest": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "position": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "bio": {
                        "type": "string"
                    },
                    "profilePic": {
                        "type": "string"
                    },
                    "linkedinUrl": {
                        "type": "string"
                    },
                    "displayOrder": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "PageResponseStaffProfileResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StaffProfileResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "StaffProfileResponse": {
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "position": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "bio": {
                        "type": "string"
                    },
                    "profilePic": {
                        "type": "string"
                    },
                    "linkedinUrl": {
                        "type": "string"
                    },
                    "displayOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    }
}